requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
pandas>=2.0.0
python-multipart>=0.0.6
torch>=2.0.0
transformers>=4.36.0
Pillow>=10.0.0
